requirements-build.txt
# Requirements for building standalone application
pyinstaller>=5.0
Pillow>=9.0.0  # For icon creation
